+++ src/service/im/v1/buzz_messages/mod.rs
@@ -1,17 +1,19 @@
-use std::collections::HashMap;
 use reqwest::Method;
 use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
 
-use crate::core::{
-    api_req::ApiRequest,
-    api_resp::{ApiResponseTrait, BaseResponse, ResponseFormat},
-    config::Config,
-    constants::AccessTokenType,
-    http::Transport,
-    req_option::RequestOption,
-    SDKResult,
+use crate::{
+    core::{
+        api_req::ApiRequest,
+        api_resp::{ApiResponseTrait, BaseResponse, ResponseFormat},
+        config::Config,
+        constants::AccessTokenType,
+        http::Transport,
+        req_option::RequestOption,
+        SDKResult,
+    },
+    service::im::v1::models::UserIdType,
 };
-use crate::service::im::v1::models::UserIdType;
 
 /// 消息加急服务
 pub struct BuzzMessagesService {
@@ -69,7 +71,10 @@ impl BuzzMessagesService {
             http_method: Method::PATCH,
             api_path: format!("/open-apis/im/v1/messages/{}/urgent_app", message_id),
             supported_access_token_types: vec![AccessTokenType::Tenant, AccessTokenType::User],
-            query_params: HashMap::from([("user_id_type".to_string(), user_id_type.as_str().to_string())]),
+            query_params: HashMap::from([(
+                "user_id_type".to_string(),
+                user_id_type.as_str().to_string(),
+            )]),
             body: serde_json::to_vec(&request)?,
             ..Default::default()
         };
@@ -89,7 +94,10 @@ impl BuzzMessagesService {
             http_method: Method::PATCH,
             api_path: format!("/open-apis/im/v1/messages/{}/urgent_sms", message_id),
             supported_access_token_types: vec![AccessTokenType::Tenant, AccessTokenType::User],
-            query_params: HashMap::from([("user_id_type".to_string(), user_id_type.as_str().to_string())]),
+            query_params: HashMap::from([(
+                "user_id_type".to_string(),
+                user_id_type.as_str().to_string(),
+            )]),
             body: serde_json::to_vec(&request)?,
             ..Default::default()
         };
@@ -109,7 +117,10 @@ impl BuzzMessagesService {
             http_method: Method::PATCH,
             api_path: format!("/open-apis/im/v1/messages/{}/urgent_phone", message_id),
             supported_access_token_types: vec![AccessTokenType::Tenant, AccessTokenType::User],
-            query_params: HashMap::from([("user_id_type".to_string(), user_id_type.as_str().to_string())]),
+            query_params: HashMap::from([(
+                "user_id_type".to_string(),
+                user_id_type.as_str().to_string(),
+            )]),
             body: serde_json::to_vec(&request)?,
             ..Default::default()
         };